+++ src/UnitTests/Providers/LinearAlgebra/Double/LinearAlgebraProviderTests.cs
@@ -358,7 +358,7 @@ namespace MathNet.Numerics.UnitTests.Providers.LinearAlgebra.Double
             var a = new double[matrix.RowCount*matrix.RowCount];
             Array.Copy(matrix.Values, a, a.Length);
 
-            var ipiv = new int[matrix.RowCount];
+            var ipiv = new long[matrix.RowCount];
 
             Control.LinearAlgebraProvider.LUFactor(a, matrix.RowCount, ipiv);
 
@@ -410,7 +410,7 @@ namespace MathNet.Numerics.UnitTests.Providers.LinearAlgebra.Double
             var a = new double[matrix.RowCount*matrix.RowCount];
             Array.Copy(matrix.Values, a, a.Length);
 
-            var ipiv = new int[matrix.RowCount];
+            var ipiv = new long[matrix.RowCount];
 
             Control.LinearAlgebraProvider.LUFactor(a, matrix.RowCount, ipiv);
             Control.LinearAlgebraProvider.LUInverseFactored(a, matrix.RowCount, ipiv);
@@ -459,7 +459,7 @@ namespace MathNet.Numerics.UnitTests.Providers.LinearAlgebra.Double
             var a = new double[matrix.RowCount*matrix.RowCount];
             Array.Copy(matrix.Values, a, a.Length);
 
-            var ipiv = new int[matrix.RowCount];
+            var ipiv = new long[matrix.RowCount];
             Control.LinearAlgebraProvider.LUFactor(a, matrix.RowCount, ipiv);
 
             var b = new[] {1.0, 2.0, 3.0, 4.0, 5.0, 6.0};
